@@ -8,7 +8,7 @@
     "watch": "webpack --watch",
     "start": "webpack-dev-server --mode development",
     "build": "webpack --mode production",
-    "api": "json-server --watch db.json --port 3004 --host 10.0.0.3"
+    "api": "json-server --watch db.json --port 3004"
   },
   "keywords": [
     "server-json"
@@ -39,8 +39,8 @@
     "rxjs": "6.5.4",
     "ts-jest": "^25.2.0",
     "ts-loader": "^6.2.1",
-    "typescript": "3.7.5",
-    "typescript-fsa": "^3.0.0-beta-2",
+    "typescript": "^3.8.3",
+    "typescript-fsa": "^3.0.0",
     "webpack": "4.41.5",
     "webpack-cli": "3.3.10",
     "webpack-dev-server": "3.10.3"
@@ -66,5 +66,8 @@
     "redux-logger": "^3.0.6",
     "redux-saga": "^1.0.0",
     "style-loader": "^1.1.3"
+  },
+  "prettier": {
+    "printWidth": 110
   }
 }
