@@ -1,23 +1,7 @@
-import { catchError, filter, map, switchMap } from "rxjs/operators";
-import { playerFetchAction } from "../typings";
-import { from, of } from "rxjs";
-import { fetchPlayerInfo } from "api/service/royalApi";
+import { from } from "rxjs";
+import { fetchPlayerInfo } from "api/service/royalApi/royal.api";
+import { coreEpic } from "utils/epic.utils";
+import { playerFetchAction } from "containers/player/typings";
 
-export const playerEpic = actions$ => {
-  return actions$.pipe(
-    filter(playerFetchAction.started.match),
-    switchMap((action: any) =>
-      from(fetchPlayerInfo("player", action.payload)).pipe(
-        map(p => playerFetchAction.done({ result: p, params: action.payload })),
-        catchError(() =>
-          of(
-            playerFetchAction.failed({
-              error: "something wrong",
-              params: action.payload
-            })
-          )
-        )
-      )
-    )
-  );
-};
+export const playerEpic = actions$ =>
+  coreEpic(actions$, playerFetchAction, payload => from(fetchPlayerInfo("players", payload)));
