@@ -11,7 +11,7 @@ import {
 import { BrowserRouter, Route } from "react-router-dom";
 import { LinkContainer } from "react-router-bootstrap";
 import { ToastContainer } from "react-toastify";
-import { Home } from "components/HomePage/home";
+import { Home } from "components/HomePage/Home";
 import { Main } from "components/main";
 import { NavRoute } from "components/typings";
 import { PlayerPage } from "components/PlayerClash/playerPage";
